Download kustomize only if binary is missing

+++ Makefile
@@ -145,10 +145,12 @@ controller-gen: ## Download controller-gen locally if necessary
 .PHONY: kustomize
 KUSTOMIZE = $(shell pwd)/bin/kustomize
 kustomize: ## Download kustomize locally if necessary
+ifeq ("$(wildcard $(KUSTOMIZE))", "")
 	rm -f ./kustomize
 	wget "https://raw.githubusercontent.com/kubernetes-sigs/kustomize/master/hack/install_kustomize.sh" -O kinstall.sh
-	chmod +x ./kinstall.sh && bash -c ./kinstall.sh && mv ./kustomize $GOBIN/kustomize
+	chmod +x ./kinstall.sh && bash -c ./kinstall.sh && mv ./kustomize $(GOBIN)/kustomize
 	rm -f ./kinstall.sh
+endif
 
 # go-get-tool will 'go install' any package $2 and install it to $1.
 PROJECT_DIR := $(shell dirname $(abspath $(lastword $(MAKEFILE_LIST))))
